refactor(api): move membership serializer
to /src/shared
