+++ api/tests/shared/unit/infrastructure/serializers/jsonapi/membership-serializer_test.js
@@ -1,6 +1,6 @@
-import { Membership } from '../../../../../lib/domain/models/Membership.js';
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/membership-serializer.js';
-import { domainBuilder, expect } from '../../../../test-helper.js';
+import { Membership } from '../../../../../../lib/domain/models/Membership.js';
+import * as serializer from '../../../../../../src/shared/infrastructure/serializers/jsonapi/membership.serializer.js';
+import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | membership-serializer', function () {
   describe('#serialize', function () {
